refactor(pi): own failure output formatting in harness
