+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Caveman style. Format: [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+- Pi failure-output normalization now lives with the Pi harness instead of the
+  harness-agnostic launch layer.
+
 ### Fixed
 - Second Pi stderr test no longer flakes under CI load: dropped the 50ms
   `_FIRST_STDOUT_AFTER_INITIAL_PROMPT_TIMEOUT_SECONDS` override from
