+++ src/meridian/lib/harness/connections/.context/FUTURE
@@ -34,9 +34,3 @@ failure path. Claude/Codex/OpenCode use bounded binary tail reads and record the
 current-launch offset. Record the stderr offset when the file is opened and read only a
 bounded tail from the current launch, decoding with replacement, then compact that.
 (Natural to fold into the managed-stdio seam above.)
-
-[pi_rpc failure-formatting boundary leak] pi_rpc.py imports compact_pi_failure_output
-from lib.launch.report (pi_rpc.py:46, 800) — Pi-specific presentation living in the
-upward, harness-agnostic launch layer. Move Pi failure normalization into a harness-owned
-module (e.g. harness/pi_failure.py or the Pi extractor surface); the launch report path
-should consume it through the existing harness bundle/extractor boundary.
